+++ multibus/__init__.py
@@ -16,7 +16,7 @@
 #
 """Multi-Bus Gateway package."""
 
-__version__ = "3.25.1"
+__version__ = "3.26.0"
 __author__ = "sm26449"
 
 from .config import Config